mcp/requirements.txt
Flask
psycopg2-binary
grpcio
grpcio-tools